# lib/api-zod/src/generated/types/blockDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zebvix Chain (ZBX) Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from './transaction';

export interface BlockDetail {
  height: number;
  hash: string;
  parentHash: string;
  timestamp: string;
  txCount: number;
  validator: string;
  gasUsed: number;
  gasLimit: number;
  size: number;
  reward: string;
  stateRoot: string;
  txHash: string;
  transactions?: Transaction[];
}